# Increase frontend build speed