--- a/.github/workflows/e2e.yml
+++ b/.github/workflows/e2e.yml
@@ -124,6 +124,15 @@ jobs:
         with:
           path: public
           key: public-cache-${{ hashFiles('pkg/webui/**', 'sdk/js/**/*.js', 'sdk/js/generated/*.json', 'config/webpack.config.babel.js', 'yarn.lock', 'sdk/js/yarn.lock')}}
+      - name: Initialize dll cache
+        if: steps.public-cache.outputs.cache-hit != 'true'
+        id: dll-cache
+        uses: actions/cache@v2
+        with:
+          path: |
+            public/libs.*.bundle.js
+            .cache/dll.json
+          key: dll-cache-${{ hashFiles('yarn.lock', 'sdk/js/yarn.lock')}}
       - name: Initialize babel cache
         id: babel-cache
         uses: actions/cache@v2
@@ -133,6 +142,9 @@ jobs:
           key: ${{ runner.os }}-babel-cache-${{ hashFiles('config/babel.config.json', 'config/webpack.config.babel.js') }}
           restore-keys: |
             ${{ runner.os }}-babel-cache-
+      - name: Build DLLs
+        if: steps.dll-cache.outputs.cache-hit != 'true'
+        run: tools/bin/mage js:buildDll
       - name: Build frontend
         if: steps.public-cache.outputs.cache-hit != 'true'
         run: tools/bin/mage js:build