--- a/config/webpack.dll.babel.js
+++ b/config/webpack.dll.babel.js
@@ -18,6 +18,9 @@ import path from 'path'
 import webpack from 'webpack'
 
 const { CONTEXT = '.', CACHE_DIR = '.cache', PUBLIC_DIR = 'public' } = process.env
+const mode = process.env.NODE_ENV === 'production' ? 'production' : 'development'
+const WEBPACK_GENERATE_PRODUCTION_SOURCEMAPS =
+  process.env.WEBPACK_GENERATE_PRODUCTION_SOURCEMAPS === 'true'
 
 const context = path.resolve(CONTEXT)
 const library = '[name]_[hash]'
@@ -25,21 +28,25 @@ const library = '[name]_[hash]'
 const pkg = require(path.resolve(context, 'package.json'))
 const excludeLibs = ['react-hot-loader', 'ttn-lw']
 const libs = Object.keys(pkg.dependencies || {}).filter(lib => !excludeLibs.includes(lib))
+const devtool =
+  (mode === 'production' && WEBPACK_GENERATE_PRODUCTION_SOURCEMAPS) || mode === 'development'
+    ? 'module-source-map'
+    : false
 
 export default {
   context,
-  mode: 'development',
+  mode,
   target: 'web',
   node: {
     fs: 'empty',
     module: 'empty',
   },
   stats: 'minimal',
-  devtool: 'module-source-map',
+  devtool,
   recordsPath: path.resolve(context, CACHE_DIR, '_libs_records'),
   entry: { libs },
   output: {
-    filename: '[name].bundle.js',
+    filename: '[name].[hash].bundle.js',
     path: path.resolve(context, PUBLIC_DIR),
     library,
   },