#### Summary
This PR will increase the frontend build speed by making use of webpacks DLL feature in production, as well as removing a calculation heavy file hashing plugin.

References #2661

```
Before ---
time mage js:clean js:cleanDeps jsSdk:Clean jsSdk:cleanDeps js:build
351.66s user 57.73s system 169% cpu 4:01.81 total

time mage js:build
91.81s user 12.12s system 119% cpu 1:27.14 total

After ---
time mage js:clean js:cleanDeps jsSdk:Clean jsSdk:cleanDeps js:build
230.47s user 33.66s system 198% cpu 2:12.87 total

time mage js:build
28.38s user 3.05s system 141% cpu 22.191 total
```
Note that these are measures from my M1 Mac, so the times on the CI machines will be significantly slower, however, the relative delta is about the same.

#### Changes
- Enable DLL referencing in production as well (currently development only)
- Enable caching of DLL module in GitHub action
- Remove `webpack-plugin-hash-output`
- Enable source map generation for release builds only
- Remove some unnecessary GitHub action job steps

#### Testing
Tested this in a separate PR #4769 

##### Regressions

In production we will now, additionally to the app bundle, also include a bundle that includes external libraries. It is possible that this leads to issues. I could however not find any when running this locally and also end-to-end tests work properly.

#### Notes for Reviewers
Note that this includes a config update, where the default values for the `js-files` config has to be updated to include the additional `libs.*.bundle.js` file that includes the external modules.

There is more speed up potential which I cannot work on currently and includes: 
- Removing number of modules (e.g. stylus files that use a lot of duplicate styles)
- Improving the mage tooling to avoid running the `yarn install` command multiple times during build
  - This could e.g. be done using yarn workspaces, so that the dependencies for webui and SDK can be installed in one go
- Treeshaking and sharing external dependencies
- Migrating to webpack 5, which (seemingly) has better build caching capabilities

#### Checklist
<!-- Make sure that this pull request is complete. -->

- [x] Scope: The referenced issue is addressed, there are no unrelated changes.
- [x] Compatibility: The changes are backwards compatible with existing API, storage, configuration and CLI, according to the compatibility commitments in `README.md` for the chosen target branch.
- [x] Documentation: Relevant documentation is added or updated.
- [x] Changelog: Significant features, behavior changes, deprecations and fixes are added to `CHANGELOG.md`.
- [x] Commits: Commit messages follow guidelines in `CONTRIBUTING.md`, there are no fixup commits left.